pyproject.toml
[project]
name = "og-webhook"
version = "0.1.0"
description = ""
authors = [
    {name = "Константин Гейдаров",email = "konstantin.geydarov@gmail.com"}
]
readme = "README.md"
requires-python = ">3.10,<4.0"
dependencies = [
    "uvicorn (>=0.35.0,<0.36.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "sshtunnel (>=0.4.0,<0.5.0)",
    "loguru (>=0.7.3,<0.8.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
